Update program.cs to get the user data and the ARD file
User Data: 
Will fetch the user data using an HTTP GET request to get the SQL data which then is received as a JSON and deserialized into a RootObject c# class that we can use.

ARD file:
Will fetch the ard file data using an HTTP GET request to get the raw string. It will then clean the string to convert it to XML and then parses it to a c# object so we can use it.